requirements.txt
feedparser
requests
beautifulsoup4
sentence_transformers
numpy
scikit-learn
matplotlib
wordcloud
umap-learn
pyzotero
nltk
tqdm
sumy
arxiv
Jinja2
openai
google-generativeai